flatten files
