--- chain/src/actors/mod.rs
@@ -1,3 +1,0 @@
-pub mod application;
-pub mod coordinator;
-pub mod indexer;
